Feature extractor
This will be the main branch for the code of the feature extractor. We will place the images needed and push merge the branch at the end. For safety reasons any modification done to this branch cannot be done directly onto it, just pull from it into your branch and change there.